Re-formatted the chat bubbles
