+++ public/stylesheets/main.css
@@ -60,3 +60,11 @@ body {
   transition: background-color 5000s ease-in-out 0s;
   -webkit-text-fill-color: #e0f2fe;
 }
+
+.chatbotTextBoxWrap input:-webkit-autofill,
+.chatbotTextBoxWrap input:-webkit-autofill:hover,
+.chatbotTextBoxWrap input:-webkit-autofill:focus,
+.chatbotTextBoxWrap input:-webkit-autofill:active {
+  transition: background-color 5000s ease-in-out 0s;
+  -webkit-text-fill-color: #e0f2fe;
+}
